Merged PR 175726: [Embed showcases] Update script consumption and add debugging options
We recently saw a problem with our showcases where a script that was hosted on a CDN would refuse connection and won't download thus causing the showcase to fail.

In this PR I changed some of the scripts to be serviced from the new playground (The others didn't exist as part of the playground, I'm working to add them to the playground and will update the showcases again once I do).

Also, some debug calls were added to the code that the playground can listen to.

*Note - The playground hosted only the un-minified version of jquery so I used that and I will update again to the minified version once my playground changes will reach production.